Swapping code
Swapped the code of both findFactorial() and reverseNumber(), to do this not only did I just copy and paste them into place, but also changed variable names / data types to make sense in the context of the functions intended use. (i.e. in reverseNumber(), factorialNumber became newNumber, vice-versa in findFactorial())